Refactor CollectionCreationRoutes
