+++ packages/frontend/src/module/collection/component/layout/MyCollectionsPageHeader/MyCollectionsPageHeader.tsx
@@ -1,7 +1,7 @@
 import MainPageHeader from "module/common/component/layout/MainPageHeader/MainPageHeader";
 import useTranslate from "module/common/hook/useTranslate";
 import { useNavigate } from "react-router-dom";
-import { CollectionRoutes } from "module/collection/CollectionRouter";
+import { CollectionRoutes } from "module/collection/router/CollectionRouter";
 import Button from "module/common/component/input/Button/Button";
 
 const MyCollectionsPageHeader = (): JSX.Element => {
