+++ packages/frontend/src/module/collection/router/CollectionRouter.tsx
@@ -1,7 +1,7 @@
 import { RouteObject } from "react-router-dom";
 import { lazy } from "react";
 import LoggedInRoute from "module/wallet/component/navigation/LoggedInRoute/LoggedInRoute";
-import useCollectionCreationRoutesObserver from "module/collection/hook/useCollectionCreationRoutesObserver";
+import CollectionCreationRoutes from "module/collection/router/CollectionCreationRoutes/CollectionCreationRoutes";
 
 export enum CollectionRoutes {
     VIEW_COLLECTION = "/collections/:id",
@@ -10,32 +10,30 @@ export enum CollectionRoutes {
     EDIT_NFT_CREATE_COLLECTION = "/collections/creation/nft/:index",
 }
 
-export const EDIT_NFT_CREATE_COLLECTION_PATH = CollectionRoutes.EDIT_NFT_CREATE_COLLECTION.replace("/:index", "");
-
-const CollectionPage = lazy(() => import("./page/CollectionPage"));
-const CollectionCreationPage = lazy(() => import("./page/CollectionCreationPage/CollectionCreationPage"));
-const EditCollectionNftDraftPage = lazy(() => import("./page/EditCollectionNftDraftPage/EditCollectionNftDraftPage"));
-const MyCollectionsPage = lazy(() => import("./page/MyCollectionsPage"));
+const CollectionPage = lazy(() => import("../page/CollectionPage"));
+const CollectionCreationPage = lazy(() => import("../page/CollectionCreationPage/CollectionCreationPage"));
+const EditCollectionNftDraftPage = lazy(() => import("../page/EditCollectionNftDraftPage/EditCollectionNftDraftPage"));
+const MyCollectionsPage = lazy(() => import("../page/MyCollectionsPage"));
 
 export const useCollectionRoutes = (): RouteObject[] => {
-    useCollectionCreationRoutesObserver();
-
     return [
         {
             path: CollectionRoutes.CREATE_COLLECTION,
             element: (
                 <LoggedInRoute>
-                    <CollectionCreationPage />
+                    <CollectionCreationRoutes />
                 </LoggedInRoute>
             ),
-        },
-        {
-            path: CollectionRoutes.EDIT_NFT_CREATE_COLLECTION,
-            element: (
-                <LoggedInRoute>
-                    <EditCollectionNftDraftPage />
-                </LoggedInRoute>
-            ),
+            children: [
+                {
+                    path: CollectionRoutes.CREATE_COLLECTION,
+                    element: <CollectionCreationPage />,
+                },
+                {
+                    path: CollectionRoutes.EDIT_NFT_CREATE_COLLECTION,
+                    element: <EditCollectionNftDraftPage />,
+                },
+            ],
         },
         {
             path: CollectionRoutes.MY_COLLECTIONS,
